Changing for ex1 and ex2
