+++ src/test/java/com/epam/tc/hw2/ex1/TestBase.java
@@ -1,28 +1,10 @@
 package com.epam.tc.hw2.ex1;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 
 public class TestBase {
 
-    public WebDriver webDriver;
-    public Elements element;
+    public static WebDriver webDriver;
+    public static Elements element;
 
-    @BeforeClass
-    public void setWebDriver() {
-        WebDriverManager.chromedriver().setup();
-        webDriver = new ChromeDriver();
-        element = new Elements(webDriver);
-        webDriver.manage().window().maximize();
-        webDriver.navigate().to(ConfProperties.getProperty("url"));
-    }
-
-    @AfterClass
-    public void closeElements() {
-        webDriver.quit();
-        element = null;
-    }
 }
