+++ src/test/java/com/epam/tc/hw2/ex2/tests/CheckEpamUserLoginTest.java
@@ -2,10 +2,23 @@ package com.epam.tc.hw2.ex2.tests;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import com.epam.tc.hw2.ex2.ConfProperties;
+import com.epam.tc.hw2.ex2.Elements;
 import com.epam.tc.hw2.ex2.TestBase;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 public class CheckEpamUserLoginTest extends TestBase {
+    @BeforeClass
+    public void setWebDriver() {
+        WebDriverManager.chromedriver().setup();
+        webDriver = new ChromeDriver();
+        element = new Elements(webDriver);
+        webDriver.manage().window().maximize();
+        webDriver.navigate().to(ConfProperties.getProperty("url"));
+    }
 
     //4. Assert Username is loggined
     @Test
