+++ src/test/java/com/epam/tc/hw2/ex2/tests/CheckSelectedElementsTest.java
@@ -4,18 +4,23 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import com.epam.tc.hw2.ex2.ConfProperties;
 import com.epam.tc.hw2.ex2.TestBase;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.Test;
 
-public class NextTest extends TestBase {
+public class CheckSelectedElementsTest extends TestBase {
 /* Assert that
 • for each checkbox there is an individual log row and value is corresponded to the status of checkbox
 • for radio button there is a log row and value is corresponded to the status of radio button
 • for dropdown there i  s a log row and value is corresponded to the selected value*/
 
+    @AfterClass
+    public void closeElements() {
+        webDriver.quit();
+        element = null;
+    }
+
     @Test
-    public void test() {
-        element.clickDropdownToggle();
-        element.login(ConfProperties.getProperty("username"), ConfProperties.getProperty("password"));
+    public void testCheckSelectedElements() {
         element.clickOnDifferentElement();
 
         element.selectCheckBoxWater();
