+++ src/test/java/resources/conf.properties
@@ -1,10 +1,5 @@
-url = https://jdi-testing.github.io/jdi-light/index.html  
+url = https://jdi-testing.github.io/jdi-light/index.html
 
 username = Roman
 password = Jdi1234
-
-logForWaterCheckBox = Colors: value changed to Yellow
-logForWindCheckBox = Wind: condition changed to true
-logForSelenRadio = metal: value changed to  Selen
-logForDropDown = Colors: value changed to Yellow
