+++ src/test/java/resources/testing/hw2.runEx1Tests.xml
@@ -1,0 +1,15 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="SuiteEx1Tests" >
+  <test name="Ex1TestsRun">
+    <classes>
+      <class name= "com.epam.tc.hw2.ex1.tests.CheckEpamUserLoginTest"/>
+      <class name= "com.epam.tc.hw2.ex1.tests.EpamPageTitleTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.EpamExistItemTitleOnTheHeaderTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.ImageExistOnTheIndexPageTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.CheckUnderIconsOnTheIndexPageTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.CheckIframeExistTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.CheckButtonIsExistOnIframeTest" />
+      <class name= "com.epam.tc.hw2.ex1.tests.CheckTitleItemsOnSideBarTest" />
+    </classes>
+  </test>
+</suite>
