+++ src/test/java/resources/testing/hw2.runEx2Tests.xml
@@ -1,0 +1,10 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="SuiteEx2Tests" >
+  <test name="Ex2TestsRun">
+    <classes>
+      <class name= "com.epam.tc.hw2.ex2.tests.CheckEpamUserLoginTest"/>
+      <class name= "com.epam.tc.hw2.ex2.tests.EpamPageTitleTest" />
+      <class name= "com.epam.tc.hw2.ex2.tests.CheckSelectedElementsTest" />
+    </classes>
+  </test>
+</suite>
