--- src/test/java/resources/testing/runEx1Tests.xml
@@ -1,8 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<suite name="SuiteEx1Tests" >
-  <test name="Ex1TestsRun">
-    <packages>
-      <package name="com.epam.tc.hw2.ex1.tests" />
-    </packages>
-  </test>
-</suite>
