--- src/test/java/resources/testing/runEx2Tests.xml
@@ -1,8 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<suite name="SuiteEx2Tests" >
-  <test name="Ex2TestsRun">
-    <packages>
-      <package name="com.epam.tc.hw2.ex2.tests" />
-    </packages>
-  </test>
-</suite>
